Python Project 2/RAG Mini Chat (Stage 2)/requirements.txt
scikit-learn
numpy
openai   # or whichever client you use
python-dotenv